LayoutTests/imported/w3c/web-platform-tests/import-maps/acquiring/modulepreload-expected.txt


Harness Error (TIMEOUT), message = null

NOTRUN After <link rel=modulepreload> import maps should fire error events
FAIL After <link rel=modulepreload> import maps are not effective assert_array_equals: expected property 0 to be "log:A" but got "log:B" (expected array ["log:A"] got ["log:B"])
